WinAPI namespace restructured
